pull logger functions out to separate file
